+++ Unity_EyapLibrary/Packages/com.maellacour.eyap-library/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Editor: Method to export all localization files as csv. Infortunately, that add the Localization package as dependency.
+
+
 ## [1.0.0] - 2023-01-05
 First release of the package.
 Contains common methods, helpers and classes to create application with Unity.